Version 0.3.52 se modifico transporte , paquete y pedido. se creo la clase pedido, con los metodos necesarios para crear uno nuevo segun el main. se corrigieron algunos typos se cambio algunos atributos de tipo de dato.

+++ TP_Amazing_part2/src/Amazing/Automovil.java
@@ -5,7 +5,7 @@ import java.util.HashMap;
 public class Automovil extends transporte {
     private int cargaMaxima;
 
-    public Automovil(String patente, double volumen, double valor, int cargaMaxima) {
+    public Automovil(String patente, int volumen, int valor, int cargaMaxima) {
         super(patente, volumen, valor); // Llama al constructor de la clase base (transporte)
         this.cargaMaxima = cargaMaxima;
     }
